pyproject.toml
[tool.poetry]
name = "embassy_bot"
version = "0.1.0"
description = ""
authors = ["oregu"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "3.11.1"
# cryptography = "^39.0.0"
asyncpg = "^0.28.0"
fastapi = "^0.101.1"
sqlalchemy = "^2.0.20"
uvicorn = "^0.23.2"
alembic = "^1.11.3"
celery = "^5.3.1"
pydantic-settings = "^2.0.3"
stringcase = "^1.2.0"
pytz = "^2023.3.post1"
pydantic = "2.1.1"
aiogram = "^3.1.1"
jinja2 = "^3.1.2"
celery-redbeat = "^2.1.1"
redis = "^5.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
